backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
opencv-python-headless
numpy
pandas
neurokit2
joblib
mediapipe
scipy
